--- a/src/Utilities/cachingoptimizer.jl
+++ b/src/Utilities/cachingoptimizer.jl
@@ -232,6 +232,24 @@ end
 
 _standardize(d::IndexMap) = d
 
+function pass_nonvariable_constraints(
+    dest::CachingOptimizer,
+    src::MOI.ModelLike,
+    idxmap::IndexMap,
+    constraint_types,
+    pass_cons = copy_constraints;
+    filter_constraints::Union{Nothing,Function} = nothing,
+)
+    dest.state == ATTACHED_OPTIMIZER && reset_optimizer(dest)
+    return pass_nonvariable_constraints(
+        dest.model_cache,
+        src,
+        idxmap,
+        constraint_types,
+        pass_cons;
+        filter_constraints = filter_constraints,
+    )
+end
 function MOI.copy_to(m::CachingOptimizer, src::MOI.ModelLike; kws...)
     if m.state == ATTACHED_OPTIMIZER
         reset_optimizer(m)