… docs/src/submodules/Utilities/overview.md
… docs/src/submodules/Utilities/reference.md
… src/Utilities/product_of_sets.jl
… src/Utilities/struct_of_constraints.jl
… src/Utilities/universalfallback.jl
… src/Utilities/vector_of_constraints.jl
… test/Utilities/copy.jl
… test/Utilities/matrix_of_constraints.jl