Current benchmark:
```julia
import MathOptInterface
const MOI = MathOptInterface
const MOIU = MOI.Utilities
import Clp
function bench(n, m)
    model = MOIU.Model{Float64}()
    x = MOI.add_variables(model, n)
    fx = MOI.SingleVariable.(x)
    ofx = 1.0 .* fx
    for i in 1:n
        MOI.add_constraint(model, ofx[i], MOI.LessThan(1.0))
        MOI.add_constraint(model, ofx[i], MOI.GreaterThan(1.0))
    end
    s = sum(ofx)
    for i in 1:n
        MOI.add_constraint(model, s, MOI.EqualTo(1.0))
    end
    optimizer = MOI.Bridges.full_bridge_optimizer(
        MOIU.CachingOptimizer(
            M(),
            Clp.Optimizer(),
        ),
        Float64,
    )
    @time MOI.copy_to(optimizer, model)
    @time MOIU.attach_optimizer(optimizer.model)
    optimizer.model
end
```

With this branch, Cbc#od/moi10 and `M=MOIU.Model{Float64}`, I get:
```
julia> bench(1000, 1000);
  0.035744 seconds (18.71 k allocations: 32.124 MiB)
  0.075348 seconds (429 allocations: 42.859 MiB)
```

With this branch, Cbc#bl/matrix and `M=Cbc.Model`, I get:
```
julia> bench(1000, 1000);
  0.011251 seconds (698 allocations: 551.617 KiB)
  0.000164 seconds (105 allocations: 146.023 KiB)
```

What really matters is the sum of the two. The speedup is approximately 10x.
~~We already have a significant decrease but I think we could have a number of allocation which does not scale linearly with `n` and `m` so I'm trying to figure it out.~~
The number of allocations is a constant + some logarithmic on `n` and `m` which is expected (e.g. because of `push` in `caches`).

See #1261

Requires #1267, #1286