@odow 2021-05-09
As another benchmark: how is the first solve? Compilation times are our big killer for this.