@blegat 2021-05-18
> And it would be better if the documentation clearly identified which functions needed to be implemented for which bits, but it may be easier to merge and keep iterating in new PRs. We still have time to break things, so we should merge and then update solvers in branches to figure out what tweaks need to be made.

Yes I agree. It's ready so that people can start play with it and we can iterate a bit before releasing.